gitlab_config/pyproject.toml
[project]
name = "gitlab-config"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "prettytable>=3.16.0",
    "python-dotenv>=1.1.0",
    "python-gitlab>=5.6.0",
    "pyyaml>=6.0.2",
    "rich>=13.0.0",
]

[dependency-groups]
dev = [
    "ruff>=0.11.9",
    "pytest>=8.0.0",
    "pytest-mock>=3.14.0",
    "isort>=6.0.1",
]

[project.scripts]
gitlab-config = "gitlab_config.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv.sources]
gitlab-config = { workspace = true }

[tool.hatch.build.targets.wheel]
packages = ["src/gitlab_config"]
